assist-final-project/requirements.txt
telethon
python-dotenv
PySocks
python-socks
certifi
